Setup limit for new cloud vaults
- create new section for cloud sync settings
- add advance cloud sync settings screen
- remove beta mode
- better UX for database options
- ability to switch between cloud and local database
- move terms and privacy into help
- support beta vaults to allow myself to continue testing restriction free

+++ src/screens/Settings/pages/DatabaseSettings/DatabaseSettings.tsx
@@ -1,4 +1,4 @@
-import { SafeAreaView, View } from 'react-native';
+import { Alert, SafeAreaView, View } from 'react-native';
 import Typography from '@/components/Typography';
 import { useStyles } from 'react-native-unistyles';
 import SettingsButton from '@/components/buttons/SettingsButton';
@@ -14,7 +14,12 @@ import { showSuccessMessage } from '@/utils/promptUtils';
 import { showErrorMessage } from '@/utils/errorUtils';
 import { Routes } from '@/navigation/Routes';
 import { useMemo } from 'react';
-import { useBoundStore } from '@/store';
+import { setCurrentDatabaseName, setResetDatabase, updateProfile, useBoundStore } from '@/store';
+import { Env } from '@/core/env';
+import { showMessage } from 'react-native-flash-message';
+import LabelButton from '@/components/buttons/LabelButton';
+import { getProfileGroup } from '@/database/supabase/getProfileGroup';
+import { syncWithSupabase } from '@/database/supabase/syncUtils';
 
 const DatabaseSettings = () => {
   const { styles } = useStyles(stylesheet);
@@ -22,15 +27,26 @@ const DatabaseSettings = () => {
 
   const handleResetDatabase = async () => {
     try {
-      if (database?.databaseName) {
-        await database.closeAsync();
-        await deleteDatabaseAsync(database?.databaseName);
-        await openDatabase({
-          currentDatabaseName: database.databaseName,
-          shouldClose: false,
-        });
-        showSuccessMessage('Database reset successfully');
-      }
+      Alert.alert('Reset Database', 'Are you sure you want to reset the database?', [
+        {
+          text: 'Cancel',
+          style: 'cancel',
+        },
+        {
+          text: 'OK',
+          onPress: async () => {
+            if (database?.databaseName) {
+              await database.closeAsync();
+              await deleteDatabaseAsync(database?.databaseName);
+              await openDatabase({
+                currentDatabaseName: database.databaseName,
+                shouldClose: false,
+              });
+              showSuccessMessage('Database reset successfully');
+            }
+          },
+        },
+      ]);
     } catch (error) {
       console.log('error', error);
       showErrorMessage('Error resetting database');
@@ -39,13 +55,25 @@ const DatabaseSettings = () => {
 
   const handleClearDatabase = async () => {
     try {
-      await dropTables(database);
-      showSuccessMessage('Database cleared successfully');
+      Alert.alert('Clear Database', 'Are you sure you want to clear the database?', [
+        {
+          text: 'Cancel',
+          style: 'cancel',
+        },
+        {
+          text: 'OK',
+          onPress: async () => {
+            await dropTables(database);
+            showSuccessMessage('Database cleared successfully');
+          },
+        },
+      ]);
     } catch (error) {
       console.log('error', error);
       showErrorMessage('Error clearing database');
     }
   };
+
   const databaseUserVersion = useMemo(
     // @ts-ignore
     () => database?.getFirstSync('PRAGMA user_version')?.user_version,
@@ -57,34 +85,113 @@ const DatabaseSettings = () => {
     navigate(Routes.DatabaseEditor);
   };
 
-  const isBetaModeEnabled = useBoundStore((state) => state.isBetaMode);
+  const showTestSettings = !__DEV__;
+  const isSyncEnabled = useBoundStore((state) => state.shouldSync);
+  const setSyncEnabled = useBoundStore((state) => state.setShouldSync);
+  const groupName = useBoundStore((state) => state.profile?.groupName);
+  const userId = useBoundStore((state) => state.profile?.id);
+
+  const handleSwitchDatabase = () => {
+    Alert.alert(
+      isSyncEnabled ? 'Switch to local vault?' : 'Switch to cloud vault?',
+      isSyncEnabled
+        ? 'Will disable cloud sync and switch to local vault.'
+        : 'Will enable cloud sync and switch to cloud vault.',
+      [
+        {
+          text: 'Cancel',
+          style: 'cancel',
+        },
+        {
+          text: 'OK',
+          onPress: async () => {
+            if (isSyncEnabled) {
+              try {
+                // setResetProfile();
+                updateProfile({
+                  groupId: Env.TEST_GROUP_ID,
+                  groupName: 'Local Vault',
+                });
+                setResetDatabase();
+                setSyncEnabled(false);
+                await openDatabase({ currentDatabaseName: Env.SQLITE_DB_NAME });
+              } catch (error) {
+                showMessage({
+                  message: 'Error',
+                  description: 'Something went wrong',
+                  type: 'danger',
+                  duration: 3000,
+                  icon: 'danger',
+                });
+              }
+            } else {
+              const { groupId, groupName } = await getProfileGroup({
+                userId,
+              });
+              if (groupId && groupName) {
+                updateProfile({
+                  groupId,
+                  groupName,
+                });
+                setSyncEnabled(true);
+                const currentDatabaseName = `${groupName}.db`;
+                setCurrentDatabaseName(currentDatabaseName);
+                await openDatabase({ currentDatabaseName });
+                await syncWithSupabase();
+              } else {
+                navigate(Routes.Login);
+              }
+            }
+          },
+        },
+      ]
+    );
+  };
 
   return (
     <SafeAreaView style={styles.container}>
       <NavBarButton style={styles.backButton} iconSource={'arrow-left'} onPress={goBack} />
 
       <View style={styles.container}>
-        <Typography variant={'titleLarge'}>Database Settings</Typography>
-        <InfoLabelButton title={'Name'} buttonTitle={database?.databaseName} />
-        <InfoLabelButton title={'Version'} buttonTitle={databaseUserVersion} />
-        <SettingsButton
-          title={'Reset database'}
-          onPress={handleResetDatabase}
-          iconSource={'safe'}
+        <Typography variant={'titleLarge'}>Database Settings.</Typography>
+        <InfoLabelButton
+          title={isSyncEnabled ? 'Cloud Vault:' : 'Local Vault:'}
+          buttonTitle={groupName || 'Local Vault'}
+          iconSource={isSyncEnabled ? 'cloud' : 'vault'}
         />
-        <SettingsButton title={'Clear database'} onPress={handleClearDatabase} iconSource={'bin'} />
-        <SettingsButton
-          title={'Open Database Details'}
-          onPress={handleNavigateToDatabaseDetails}
-          iconSource={'settings'}
+        <LabelButton
+          title={isSyncEnabled ? 'Switch to local vault' : 'Switch to cloud vault'}
+          onPress={handleSwitchDatabase}
         />
-        {isBetaModeEnabled && (
+        {showTestSettings && (
+          <InfoLabelButton title={'Version'} buttonTitle={databaseUserVersion} />
+        )}
+
+        {showTestSettings && (
           <SettingsButton
-            title={'Sync Database'}
-            onPress={() => navigate(Routes.SyncSettings)}
-            iconSource={'cloud'}
+            title={'Open Database Details'}
+            onPress={handleNavigateToDatabaseDetails}
+            iconSource={'settings'}
           />
         )}
+        <View style={styles.dangerZoneContainer}>
+          <Typography variant={'titleLarge'} style={styles.dangerZoneTitle}>
+            Danger Zone:
+          </Typography>
+          <Typography variant={'bodyMedium'}>
+            Having trouble with the app? Try these options:
+          </Typography>
+          <SettingsButton
+            title={'Clear All Recipes'}
+            onPress={handleClearDatabase}
+            iconSource={'bin'}
+          />
+          <SettingsButton
+            title={'Delete and Reset Database'}
+            onPress={handleResetDatabase}
+            iconSource={'bin'}
+          />
+        </View>
       </View>
     </SafeAreaView>
   );
